WOO-HOO! BACK TO THE MAIN BRANCH!
This version of APBS is stable and should be considered a 0.2.2 release
candidate.  It includes some restructuring of Vacc and Vpmg as well as
considerable optimization of the force evaluation routines.  It does not,
however, include much optimization based on profiling -- work that still needs
to be done before 0.2.2 gets released.